+++ src/app/modules/showcase-ui/examples/alerts-example/alerts-example.component.ts
@@ -1,9 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { Alert, AlertComponent } from '@lab900/ui';
 
 @Component({
   selector: 'lab900-alerts-example',
   imports: [AlertComponent],
+  // eslint-disable-next-line
+  changeDetection: ChangeDetectionStrategy.Eager,
   template: `
     <lab900-alert [type]="Error">This is an error!</lab900-alert>
     <br />