+++ src/app/modules/showcase-ui/examples/merger-example/custom-example.component.ts
@@ -1,10 +1,10 @@
-import { Component, Input } from '@angular/core';
+import { Component, input } from '@angular/core';
 import { MergerItemComponent } from '@lab900/ui';
 import { MergerDataExample } from './models/merger-data-example.model';
 
 @Component({
   selector: 'lab900-custom-component-example',
-  template: ` @if (data) {
+  template: `@if (data(); as data) {
     <div>
       <p>{{ data.text }}</p>
     </div>
@@ -12,6 +12,5 @@ import { MergerDataExample } from './models/merger-data-example.model';
   imports: [],
 })
 export class CustomExampleComponent implements MergerItemComponent<MergerDataExample> {
-  @Input()
-  public data?: MergerDataExample;
+  public readonly data = input<MergerDataExample>();
 }